fix bugs and structures
