+++ .prettierrc.json
@@ -1,0 +1,10 @@
+{
+  "useTabs": true,
+  "tabWidth": 2,
+  "printWidth": 90,
+  "semi": true,
+  "trailingComma": "es5",
+  "singleQuote": true,
+  "bracketSpacing": true,
+  "endOfLine": "auto"
+}
