--- src/Lists/Categories.tsx
@@ -1,18 +1,0 @@
-/* VENDOR */
-import { useSelector } from "react-redux";
-
-/* APPLICATION */
-import { ListItem } from "./ListItem";
-import { selectAllCategories } from "../features/categoriesSlice";
-
-export const Categories = () => {
-  const categories = useSelector(selectAllCategories);
-
-  return (
-    <ul>
-      {categories.map((category) => (
-        <ListItem key={category.id} item={category} />
-      ))}
-    </ul>
-  );
-};
