--- src/Modal/ModalText.tsx
@@ -1,7 +1,0 @@
-interface ModalTextProps {
-  text: string;
-}
-
-export const ModalText: React.FC<ModalTextProps> = ({ text }) => {
-  return <p className="modal__content-text">{text}</p>;
-};
