--- src/Modal/ModalTextarea.tsx
@@ -1,22 +1,0 @@
-interface ModalTextareaProps {
-  description: string;
-  setDescription: React.Dispatch<React.SetStateAction<string>>;
-}
-
-export const ModalTextarea: React.FC<ModalTextareaProps> = ({
-  description,
-  setDescription,
-}) => {
-  return (
-    <div className="modaltextarea-wrapper">
-      <label htmlFor="modaltextarea">Описание</label>
-      <textarea
-        id="modaltextarea"
-        className="modaltextarea"
-        placeholder="Введите описание задачи"
-        value={description}
-        onChange={(e) => setDescription(e.target.value)}
-      />
-    </div>
-  );
-};
