+++ src/components/Categories/Categories.tsx
@@ -1,0 +1,16 @@
+/* APPLICATION */
+import { ListItem } from '../ListItem/ListItem';
+import { selectAllCategories } from '../../features/categoriesSlice';
+import { useAppSelector } from '../../app/hooks';
+
+export const Categories = () => {
+	const categories = useAppSelector(selectAllCategories);
+
+	return (
+		<ul>
+			{categories.map((category) => (
+				<ListItem key={category.id} item={category} />
+			))}
+		</ul>
+	);
+};
