+++ src/components/ListItem/types.ts
@@ -1,0 +1,8 @@
+export interface ListItemProps {
+	item: {
+		id: string;
+		name: string;
+		description: string;
+		category?: string;
+	};
+}
