+++ src/components/ModalBtn/types.ts
@@ -1,0 +1,8 @@
+import type { ReactNode } from 'react';
+
+export interface ModalBtnProps {
+	type?: string;
+	children: ReactNode;
+	size?: string;
+	onClick: () => void;
+}
