+++ src/components/ModalCreateItem/types.ts
@@ -1,0 +1,6 @@
+import { Dispatch, SetStateAction } from 'react';
+
+export interface ModalCreateItemProps {
+	active: boolean;
+	setActive: Dispatch<SetStateAction<boolean>>;
+}
