+++ src/components/ModalDropdown/types.ts
@@ -1,0 +1,6 @@
+import { Dispatch, SetStateAction } from 'react';
+
+export interface ModalDropdownProps {
+	selected: string | undefined;
+	setSelected: Dispatch<SetStateAction<string>>;
+}
