+++ src/components/ModalFooter/types.ts
@@ -1,0 +1,9 @@
+import { Dispatch, SetStateAction } from 'react';
+
+export interface ModalFooterProps {
+	clearState?: () => void;
+	setActive: Dispatch<SetStateAction<boolean>>;
+	submitBtnText: string;
+	size?: string;
+	onSubmit: () => void;
+}
