+++ src/components/ModalHeader/types.ts
@@ -1,0 +1,7 @@
+import { Dispatch, SetStateAction } from 'react';
+
+export interface ModalHeaderProps {
+	clearState?: () => void;
+	setActive: Dispatch<SetStateAction<boolean>>;
+	title: string;
+}
