+++ src/components/ModalInput/types.ts
@@ -1,0 +1,7 @@
+import { Dispatch, SetStateAction } from 'react';
+
+export interface ModalInputProps {
+	name: string;
+	setName: Dispatch<SetStateAction<string>>;
+	size?: string;
+}
