+++ src/components/ModalRow/types.ts
@@ -1,0 +1,8 @@
+import { Dispatch, SetStateAction } from 'react';
+
+export interface ModalRowProps {
+	name: string;
+	setName: Dispatch<SetStateAction<string>>;
+	selected: string | undefined;
+	setSelected: Dispatch<SetStateAction<string>>;
+}
