+++ src/components/ModalText/ModalText.tsx
@@ -1,0 +1,6 @@
+import React, { type FC } from 'react';
+import { ModalTextProps } from './types';
+
+export const ModalText: FC<ModalTextProps> = ({ text }) => {
+	return <p className="modal__content-text">{text}</p>;
+};
