+++ src/components/ModalText/types.ts
@@ -1,0 +1,3 @@
+export interface ModalTextProps {
+	text: string;
+}
