+++ src/components/ModalTextarea/types.ts
@@ -1,0 +1,6 @@
+import { Dispatch, SetStateAction } from 'react';
+
+export interface ModalTextareaProps {
+	description: string;
+	setDescription: Dispatch<SetStateAction<string>>;
+}
