+++ src/features/tasksSlice.ts
@@ -1,5 +1,5 @@
 /* VENDOR */
-import { createSlice, PayloadAction } from "@reduxjs/toolkit";
+import { createSlice } from "@reduxjs/toolkit";
 import { v4 as uuidv4 } from "uuid";
 
 /* APPLICATION */
@@ -61,9 +61,9 @@ export const tasksSlice = createSlice({
       state.splice(rmTaskIndex, 1);
     },
     tasksClearedCategories: (state, action) => {
-      state.map((task) => {
+      state.forEach((task) => {
         if (task.category === action.payload) task.category = "";
-      });
+      } );
     },
   },
 });
